lesson_3_6_Tests_нужна помощь по коду в тестовой части,исправлен контроллер Faculty

+++ src/main/java/ru/hogwarts/school/controller/FacultyController.java
@@ -18,13 +18,11 @@ public class FacultyController {
         this.facultyService = facultyService;
     }
 
-    // запись в БД появляется, но с непонятно места id, от чего это зависит?
     @PostMapping//POST
     public Faculty addFaculty(@RequestBody Faculty faculty) {
         return facultyService.addFaculty(faculty);
     }
 
-    //  почему не видит данные из БД, не могу получить ранее созданные. только последний 12
     @GetMapping("{id}")//GET
     public ResponseEntity<Faculty> getFacultyInfo(@PathVariable long id) {
         Faculty faculty = facultyService.findFaculty(id);
@@ -33,7 +31,6 @@ public class FacultyController {
         }
         return ResponseEntity.ok(faculty);
     }
-//создается новая строка что плохо как убрать через проверку наличия ID?
 
     @PutMapping//PUT
     public ResponseEntity<Faculty> editFaculty(@RequestBody Faculty faculty) {
@@ -43,11 +40,13 @@ public class FacultyController {
         }
         return ResponseEntity.ok(foundFaculty);
     }
-    // ошибка 500, как ее избежать?
-    //
-    //org.springframework.dao.EmptyResultDataAccessException: No class ru.hogwarts.school.model.Faculty entity with id 0 exists!
+
     @DeleteMapping("{id}")//DELETE
     public ResponseEntity<Faculty> deleteFaculty(@PathVariable long id) {
+        Faculty faculty = facultyService.findFaculty(id);
+        if (faculty == null) {
+            return ResponseEntity.notFound().build();
+        }
         facultyService.deleteFaculty(id);
         return ResponseEntity.ok().build();
     }
